+++ Assets/Scripts/HexGrid.cs
@@ -7,8 +7,8 @@ using System.Collections.Generic;
 /// <summary>
 /// Component that represents an entire hexagon map.
 /// </summary>
-public class HexGrid : MonoBehaviour {
-
+public class HexGrid : MonoBehaviour
+{
 	[SerializeField]
 	HexCell cellPrefab;
 
@@ -30,12 +30,14 @@ public class HexGrid : MonoBehaviour {
 	/// <summary>
 	/// Amount of cells in the X dimension.
 	/// </summary>
-	public int CellCountX { get; private set; }
+	public int CellCountX
+	{ get; private set; }
 
 	/// <summary>
 	/// Amount of cells in the Z dimension.
 	/// </summary>
-	public int CellCountZ { get; private set; }
+	public int CellCountZ
+	{ get; private set; }
 
 	/// <summary>
 	/// Whether there currently exists a path that should be displayed.
@@ -45,7 +47,8 @@ public class HexGrid : MonoBehaviour {
 	/// <summary>
 	/// Whether east-west wrapping is enabled.
 	/// </summary>
-	public bool Wrapping { get; private set; }
+	public bool Wrapping
+	{ get; private set; }
 
 	Transform[] columns;
 	HexGridChunk[] chunks;
@@ -66,7 +69,8 @@ public class HexGrid : MonoBehaviour {
 
 	HexCellShaderData cellShaderData;
 
-	void Awake () {
+	void Awake ()
+	{
 		CellCountX = 20;
 		CellCountZ = 15;
 		HexMetrics.noiseSource = noiseSource;
@@ -83,7 +87,8 @@ public class HexGrid : MonoBehaviour {
 	/// <param name="unit">Unit to add.</param>
 	/// <param name="location">Cell in which to place the unit.</param>
 	/// <param name="orientation">Orientation of the unit.</param>
-	public void AddUnit (HexUnit unit, HexCell location, float orientation) {
+	public void AddUnit (HexUnit unit, HexCell location, float orientation)
+	{
 		units.Add(unit);
 		unit.Grid = this;
 		unit.Location = location;
@@ -94,7 +99,8 @@ public class HexGrid : MonoBehaviour {
 	/// Remove a unit from the map.
 	/// </summary>
 	/// <param name="unit">The unit to remove.</param>
-	public void RemoveUnit (HexUnit unit) {
+	public void RemoveUnit (HexUnit unit)
+	{
 		units.Remove(unit);
 		unit.Die();
 	}
@@ -115,19 +121,23 @@ public class HexGrid : MonoBehaviour {
 	/// <param name="wrapping">Whether the map wraps east-west.</param>
 	/// <returns>Whether the map was successfully created. It fails if the X or Z size
 	/// is not a multiple of the respective chunk size.</returns>
-	public bool CreateMap (int x, int z, bool wrapping) {
+	public bool CreateMap (int x, int z, bool wrapping)
+	{
 		if (
 			x <= 0 || x % HexMetrics.chunkSizeX != 0 ||
 			z <= 0 || z % HexMetrics.chunkSizeZ != 0
-		) {
+		)
+		{
 			Debug.LogError("Unsupported map size.");
 			return false;
 		}
 
 		ClearPath();
 		ClearUnits();
-		if (columns != null) {
-			for (int i = 0; i < columns.Length; i++) {
+		if (columns != null)
+		{
+			for (int i = 0; i < columns.Length; i++)
+			{
 				Destroy(columns[i].gameObject);
 			}
 		}
@@ -145,41 +155,52 @@ public class HexGrid : MonoBehaviour {
 		return true;
 	}
 
-	void CreateChunks () {
+	void CreateChunks ()
+	{
 		columns = new Transform[chunkCountX];
-		for (int x = 0; x < chunkCountX; x++) {
+		for (int x = 0; x < chunkCountX; x++)
+		{
 			columns[x] = new GameObject("Column").transform;
 			columns[x].SetParent(transform, false);
 		}
 
 		chunks = new HexGridChunk[chunkCountX * chunkCountZ];
-		for (int z = 0, i = 0; z < chunkCountZ; z++) {
-			for (int x = 0; x < chunkCountX; x++) {
+		for (int z = 0, i = 0; z < chunkCountZ; z++)
+		{
+			for (int x = 0; x < chunkCountX; x++)
+			{
 				HexGridChunk chunk = chunks[i++] = Instantiate(chunkPrefab);
 				chunk.transform.SetParent(columns[x], false);
 			}
 		}
 	}
 
-	void CreateCells () {
+	void CreateCells ()
+	{
 		cells = new HexCell[CellCountZ * CellCountX];
 
-		for (int z = 0, i = 0; z < CellCountZ; z++) {
-			for (int x = 0; x < CellCountX; x++) {
+		for (int z = 0, i = 0; z < CellCountZ; z++)
+		{
+			for (int x = 0; x < CellCountX; x++)
+			{
 				CreateCell(x, z, i++);
 			}
 		}
 	}
 
-	void ClearUnits () {
-		for (int i = 0; i < units.Count; i++) {
+	void ClearUnits ()
+	{
+		for (int i = 0; i < units.Count; i++)
+		{
 			units[i].Die();
 		}
 		units.Clear();
 	}
 
-	void OnEnable () {
-		if (!HexMetrics.noiseSource) {
+	void OnEnable ()
+	{
+		if (!HexMetrics.noiseSource)
+		{
 			HexMetrics.noiseSource = noiseSource;
 			HexMetrics.InitializeHashGrid(seed);
 			HexUnit.unitPrefab = unitPrefab;
@@ -193,9 +214,11 @@ public class HexGrid : MonoBehaviour {
 	/// </summary>
 	/// <param name="ray"><see cref="Ray"/> used to perform a raycast.</param>
 	/// <returns>The hit cell, if any.</returns>
-	public HexCell GetCell (Ray ray) {
+	public HexCell GetCell (Ray ray)
+	{
 		RaycastHit hit;
-		if (Physics.Raycast(ray, out hit)) {
+		if (Physics.Raycast(ray, out hit))
+		{
 			return GetCell(hit.point);
 		}
 		return null;
@@ -206,7 +229,8 @@ public class HexGrid : MonoBehaviour {
 	/// </summary>
 	/// <param name="position">Position to check.</param>
 	/// <returns>The cell containing the position, if it exists.</returns>
-	public HexCell GetCell (Vector3 position) {
+	public HexCell GetCell (Vector3 position)
+	{
 		position = transform.InverseTransformPoint(position);
 		HexCoordinates coordinates = HexCoordinates.FromPosition(position);
 		return GetCell(coordinates);
@@ -217,16 +241,34 @@ public class HexGrid : MonoBehaviour {
 	/// </summary>
 	/// <param name="coordinates"><see cref="HexCoordinates"/> of the cell.</param>
 	/// <returns>The cell with the given coordinates, if it exists.</returns>
-	public HexCell GetCell (HexCoordinates coordinates) {
+	public HexCell GetCell (HexCoordinates coordinates)
+	{
 		int z = coordinates.Z;
-		if (z < 0 || z >= CellCountZ) {
-			return null;
-		}
 		int x = coordinates.X + z / 2;
-		if (x < 0 || x >= CellCountX) {
+		if (z < 0 || z >= CellCountZ || x < 0 || x >= CellCountX)
+		{
 			return null;
 		}
 		return cells[x + z * CellCountX];
+	}
+
+	/// <summary>
+	/// Try to get the cell with specific <see cref="HexCoordinates"/>.
+	/// </summary>
+	/// <param name="coordinates"><see cref="HexCoordinates"/> of the cell.</param>
+	/// <param name="cell">The cell, if it exists.</param>
+	/// <returns>Whether the cell exists.</returns>
+	public bool TryGetCell (HexCoordinates coordinates, out HexCell cell)
+	{
+		int z = coordinates.Z;
+		int x = coordinates.X + z / 2;
+		if (z < 0 || z >= CellCountZ || x < 0 || x >= CellCountX)
+		{
+			cell = null;
+			return false;
+		}
+		cell = cells[x + z * CellCountX];
+		return true;
 	}
 
 	/// <summary>
@@ -249,60 +291,37 @@ public class HexGrid : MonoBehaviour {
 	/// Control whether the map UI should be visible or hidden.
 	/// </summary>
 	/// <param name="visible">Whether the UI should be visibile.</param>
-	public void ShowUI (bool visible) {
-		for (int i = 0; i < chunks.Length; i++) {
+	public void ShowUI (bool visible)
+	{
+		for (int i = 0; i < chunks.Length; i++)
+		{
 			chunks[i].ShowUI(visible);
 		}
 	}
 
-	void CreateCell (int x, int z, int i) {
+	void CreateCell (int x, int z, int i)
+	{
 		Vector3 position;
 		position.x = (x + z * 0.5f - z / 2) * HexMetrics.innerDiameter;
 		position.y = 0f;
 		position.z = z * (HexMetrics.outerRadius * 1.5f);
 
 		HexCell cell = cells[i] = Instantiate<HexCell>(cellPrefab);
+		cell.Grid = this;
 		cell.transform.localPosition = position;
 		cell.Coordinates = HexCoordinates.FromOffsetCoordinates(x, z);
 		cell.Index = i;
 		cell.ColumnIndex = x / HexMetrics.chunkSizeX;
 		cell.ShaderData = cellShaderData;
 
-		if (Wrapping) {
+		if (Wrapping)
+		{
 			cell.Explorable = z > 0 && z < CellCountZ - 1;
 		}
-		else {
+		else
+		{
 			cell.Explorable =
 				x > 0 && z > 0 && x < CellCountX - 1 && z < CellCountZ - 1;
-		}
-
-		if (x > 0) {
-			cell.SetNeighbor(HexDirection.W, cells[i - 1]);
-			if (Wrapping && x == CellCountX - 1) {
-				cell.SetNeighbor(HexDirection.E, cells[i - x]);
-			}
-		}
-		if (z > 0) {
-			if ((z & 1) == 0) {
-				cell.SetNeighbor(HexDirection.SE, cells[i - CellCountX]);
-				if (x > 0) {
-					cell.SetNeighbor(HexDirection.SW, cells[i - CellCountX - 1]);
-				}
-				else if (Wrapping) {
-					cell.SetNeighbor(HexDirection.SW, cells[i - 1]);
-				}
-			}
-			else {
-				cell.SetNeighbor(HexDirection.SW, cells[i - CellCountX]);
-				if (x < CellCountX - 1) {
-					cell.SetNeighbor(HexDirection.SE, cells[i - CellCountX + 1]);
-				}
-				else if (Wrapping) {
-					cell.SetNeighbor(
-						HexDirection.SE, cells[i - CellCountX * 2 + 1]
-					);
-				}
-			}
 		}
 
 		Text label = Instantiate<Text>(cellLabelPrefab);
@@ -315,7 +334,8 @@ public class HexGrid : MonoBehaviour {
 		AddCellToChunk(x, z, cell);
 	}
 
-	void AddCellToChunk (int x, int z, HexCell cell) {
+	void AddCellToChunk (int x, int z, HexCell cell)
+	{
 		int chunkX = x / HexMetrics.chunkSizeX;
 		int chunkZ = z / HexMetrics.chunkSizeZ;
 		HexGridChunk chunk = chunks[chunkX + chunkZ * chunkCountX];
@@ -329,17 +349,20 @@ public class HexGrid : MonoBehaviour {
 	/// Save the map.
 	/// </summary>
 	/// <param name="writer"><see cref="BinaryWriter"/> to use.</param>
-	public void Save (BinaryWriter writer) {
+	public void Save (BinaryWriter writer)
+	{
 		writer.Write(CellCountX);
 		writer.Write(CellCountZ);
 		writer.Write(Wrapping);
 
-		for (int i = 0; i < cells.Length; i++) {
+		for (int i = 0; i < cells.Length; i++)
+		{
 			cells[i].Save(writer);
 		}
 
 		writer.Write(units.Count);
-		for (int i = 0; i < units.Count; i++) {
+		for (int i = 0; i < units.Count; i++)
+		{
 			units[i].Save(writer);
 		}
 	}
@@ -349,17 +372,21 @@ public class HexGrid : MonoBehaviour {
 	/// </summary>
 	/// <param name="reader"><see cref="BinaryReader"/> to use.</param>
 	/// <param name="header">Header version.</param>
-	public void Load (BinaryReader reader, int header) {
+	public void Load (BinaryReader reader, int header)
+	{
 		ClearPath();
 		ClearUnits();
 		int x = 20, z = 15;
-		if (header >= 1) {
+		if (header >= 1)
+		{
 			x = reader.ReadInt32();
 			z = reader.ReadInt32();
 		}
 		bool wrapping = header >= 5 ? reader.ReadBoolean() : false;
-		if (x != CellCountX || z != CellCountZ || this.Wrapping != wrapping) {
-			if (!CreateMap(x, z, wrapping)) {
+		if (x != CellCountX || z != CellCountZ || this.Wrapping != wrapping)
+		{
+			if (!CreateMap(x, z, wrapping))
+			{
 				return;
 			}
 		}
@@ -367,16 +394,20 @@ public class HexGrid : MonoBehaviour {
 		bool originalImmediateMode = cellShaderData.ImmediateMode;
 		cellShaderData.ImmediateMode = true;
 
-		for (int i = 0; i < cells.Length; i++) {
+		for (int i = 0; i < cells.Length; i++)
+		{
 			cells[i].Load(reader, header);
 		}
-		for (int i = 0; i < chunks.Length; i++) {
+		for (int i = 0; i < chunks.Length; i++)
+		{
 			chunks[i].Refresh();
 		}
 
-		if (header >= 2) {
+		if (header >= 2)
+		{
 			int unitCount = reader.ReadInt32();
-			for (int i = 0; i < unitCount; i++) {
+			for (int i = 0; i < unitCount; i++)
+			{
 				HexUnit.Load(reader, this);
 			}
 		}
@@ -388,12 +419,15 @@ public class HexGrid : MonoBehaviour {
 	/// Get a list of cells representing the currently visible path.
 	/// </summary>
 	/// <returns>The current path list, if a visible path exists.</returns>
-	public List<HexCell> GetPath () {
-		if (!currentPathExists) {
+	public List<HexCell> GetPath ()
+	{
+		if (!currentPathExists)
+		{
 			return null;
 		}
 		List<HexCell> path = ListPool<HexCell>.Get();
-		for (HexCell c = currentPathTo; c != currentPathFrom; c = c.PathFrom) {
+		for (HexCell c = currentPathTo; c != currentPathFrom; c = c.PathFrom)
+		{
 			path.Add(c);
 		}
 		path.Add(currentPathFrom);
@@ -404,10 +438,13 @@ public class HexGrid : MonoBehaviour {
 	/// <summary>
 	/// Clear the current path.
 	/// </summary>
-	public void ClearPath () {
-		if (currentPathExists) {
+	public void ClearPath ()
+	{
+		if (currentPathExists)
+		{
 			HexCell current = currentPathTo;
-			while (current != currentPathFrom) {
+			while (current != currentPathFrom)
+			{
 				current.SetLabel(null);
 				current.DisableHighlight();
 				current = current.PathFrom;
@@ -415,17 +452,21 @@ public class HexGrid : MonoBehaviour {
 			current.DisableHighlight();
 			currentPathExists = false;
 		}
-		else if (currentPathFrom) {
+		else if (currentPathFrom)
+		{
 			currentPathFrom.DisableHighlight();
 			currentPathTo.DisableHighlight();
 		}
 		currentPathFrom = currentPathTo = null;
 	}
 
-	void ShowPath (int speed) {
-		if (currentPathExists) {
+	void ShowPath (int speed)
+	{
+		if (currentPathExists)
+		{
 			HexCell current = currentPathTo;
-			while (current != currentPathFrom) {
+			while (current != currentPathFrom)
+			{
 				int turn = (current.Distance - 1) / speed;
 				current.SetLabel(turn.ToString());
 				current.EnableHighlight(Color.white);
@@ -442,7 +483,8 @@ public class HexGrid : MonoBehaviour {
 	/// <param name="fromCell">Cell to start the search from.</param>
 	/// <param name="toCell">Cell to find a path towards.</param>
 	/// <param name="unit">Unit for which the path is.</param>
-	public void FindPath (HexCell fromCell, HexCell toCell, HexUnit unit) {
+	public void FindPath (HexCell fromCell, HexCell toCell, HexUnit unit)
+	{
 		ClearPath();
 		currentPathFrom = fromCell;
 		currentPathTo = toCell;
@@ -450,52 +492,62 @@ public class HexGrid : MonoBehaviour {
 		ShowPath(unit.Speed);
 	}
 
-	bool Search (HexCell fromCell, HexCell toCell, HexUnit unit) {
+	bool Search (HexCell fromCell, HexCell toCell, HexUnit unit)
+	{
 		int speed = unit.Speed;
 		searchFrontierPhase += 2;
-		if (searchFrontier == null) {
+		if (searchFrontier == null)
+		{
 			searchFrontier = new HexCellPriorityQueue();
 		}
-		else {
+		else
+		{
 			searchFrontier.Clear();
 		}
 
 		fromCell.SearchPhase = searchFrontierPhase;
 		fromCell.Distance = 0;
 		searchFrontier.Enqueue(fromCell);
-		while (searchFrontier.Count > 0) {
+		while (searchFrontier.Count > 0)
+		{
 			HexCell current = searchFrontier.Dequeue();
 			current.SearchPhase += 1;
 
-			if (current == toCell) {
+			if (current == toCell)
+			{
 				return true;
 			}
 
 			int currentTurn = (current.Distance - 1) / speed;
 
-			for (HexDirection d = HexDirection.NE; d <= HexDirection.NW; d++) {
-				HexCell neighbor = current.GetNeighbor(d);
+			for (HexDirection d = HexDirection.NE; d <= HexDirection.NW; d++)
+			{
 				if (
-					neighbor == null ||
+					!current.TryGetNeighbor(d, out HexCell neighbor) ||
 					neighbor.SearchPhase > searchFrontierPhase
-				) {
+				)
+				{
 					continue;
 				}
-				if (!unit.IsValidDestination(neighbor)) {
+				if (!unit.IsValidDestination(neighbor))
+				{
 					continue;
 				}
 				int moveCost = unit.GetMoveCost(current, neighbor, d);
-				if (moveCost < 0) {
+				if (moveCost < 0)
+				{
 					continue;
 				}
 
 				int distance = current.Distance + moveCost;
 				int turn = (distance - 1) / speed;
-				if (turn > currentTurn) {
+				if (turn > currentTurn)
+				{
 					distance = turn * speed + moveCost;
 				}
 
-				if (neighbor.SearchPhase < searchFrontierPhase) {
+				if (neighbor.SearchPhase < searchFrontierPhase)
+				{
 					neighbor.SearchPhase = searchFrontierPhase;
 					neighbor.Distance = distance;
 					neighbor.PathFrom = current;
@@ -503,7 +555,8 @@ public class HexGrid : MonoBehaviour {
 						neighbor.Coordinates.DistanceTo(toCell.Coordinates);
 					searchFrontier.Enqueue(neighbor);
 				}
-				else if (distance < neighbor.Distance) {
+				else if (distance < neighbor.Distance)
+				{
 					int oldPriority = neighbor.SearchPriority;
 					neighbor.Distance = distance;
 					neighbor.PathFrom = current;
@@ -519,9 +572,11 @@ public class HexGrid : MonoBehaviour {
 	/// </summary>
 	/// <param name="fromCell">Cell from which to start viewing.</param>
 	/// <param name="range">Visibility range.</param>
-	public void IncreaseVisibility (HexCell fromCell, int range) {
+	public void IncreaseVisibility (HexCell fromCell, int range)
+	{
 		List<HexCell> cells = GetVisibleCells(fromCell, range);
-		for (int i = 0; i < cells.Count; i++) {
+		for (int i = 0; i < cells.Count; i++)
+		{
 			cells[i].IncreaseVisibility();
 		}
 		ListPool<HexCell>.Add(cells);
@@ -532,9 +587,11 @@ public class HexGrid : MonoBehaviour {
 	/// </summary>
 	/// <param name="fromCell">Cell from which to stop viewing.</param>
 	/// <param name="range">Visibility range.</param>
-	public void DecreaseVisibility (HexCell fromCell, int range) {
+	public void DecreaseVisibility (HexCell fromCell, int range)
+	{
 		List<HexCell> cells = GetVisibleCells(fromCell, range);
-		for (int i = 0; i < cells.Count; i++) {
+		for (int i = 0; i < cells.Count; i++)
+		{
 			cells[i].DecreaseVisibility();
 		}
 		ListPool<HexCell>.Add(cells);
@@ -544,23 +601,28 @@ public class HexGrid : MonoBehaviour {
 	/// Reset visibility of the entire map, viewing from all units.
 	/// </summary>
 	public void ResetVisibility () {
-		for (int i = 0; i < cells.Length; i++) {
+		for (int i = 0; i < cells.Length; i++)
+		{
 			cells[i].ResetVisibility();
 		}
-		for (int i = 0; i < units.Count; i++) {
+		for (int i = 0; i < units.Count; i++)
+		{
 			HexUnit unit = units[i];
 			IncreaseVisibility(unit.Location, unit.VisionRange);
 		}
 	}
 
-	List<HexCell> GetVisibleCells (HexCell fromCell, int range) {
+	List<HexCell> GetVisibleCells (HexCell fromCell, int range)
+	{
 		List<HexCell> visibleCells = ListPool<HexCell>.Get();
 
 		searchFrontierPhase += 2;
-		if (searchFrontier == null) {
+		if (searchFrontier == null)
+		{
 			searchFrontier = new HexCellPriorityQueue();
 		}
-		else {
+		else
+		{
 			searchFrontier.Clear();
 		}
 
@@ -569,35 +631,40 @@ public class HexGrid : MonoBehaviour {
 		fromCell.Distance = 0;
 		searchFrontier.Enqueue(fromCell);
 		HexCoordinates fromCoordinates = fromCell.Coordinates;
-		while (searchFrontier.Count > 0) {
+		while (searchFrontier.Count > 0)
+		{
 			HexCell current = searchFrontier.Dequeue();
 			current.SearchPhase += 1;
 			visibleCells.Add(current);
 
-			for (HexDirection d = HexDirection.NE; d <= HexDirection.NW; d++) {
-				HexCell neighbor = current.GetNeighbor(d);
+			for (HexDirection d = HexDirection.NE; d <= HexDirection.NW; d++)
+			{
 				if (
-					neighbor == null ||
+					!current.TryGetNeighbor(d, out HexCell neighbor) ||
 					neighbor.SearchPhase > searchFrontierPhase ||
 					!neighbor.Explorable
-				) {
+				)
+				{
 					continue;
 				}
 
 				int distance = current.Distance + 1;
 				if (distance + neighbor.ViewElevation > range ||
 					distance > fromCoordinates.DistanceTo(neighbor.Coordinates)
-				) {
+				)
+				{
 					continue;
 				}
 
-				if (neighbor.SearchPhase < searchFrontierPhase) {
+				if (neighbor.SearchPhase < searchFrontierPhase)
+				{
 					neighbor.SearchPhase = searchFrontierPhase;
 					neighbor.Distance = distance;
 					neighbor.SearchHeuristic = 0;
 					searchFrontier.Enqueue(neighbor);
 				}
-				else if (distance < neighbor.Distance) {
+				else if (distance < neighbor.Distance)
+				{
 					int oldPriority = neighbor.SearchPriority;
 					neighbor.Distance = distance;
 					searchFrontier.Change(neighbor, oldPriority);
@@ -611,11 +678,13 @@ public class HexGrid : MonoBehaviour {
 	/// Center the map given an X position, to facilitate east-west wrapping.
 	/// </summary>
 	/// <param name="xPosition">X position.</param>
-	public void CenterMap (float xPosition) {
+	public void CenterMap (float xPosition)
+	{
 		int centerColumnIndex = (int)
 			(xPosition / (HexMetrics.innerDiameter * HexMetrics.chunkSizeX));
 		
-		if (centerColumnIndex == currentCenterColumnIndex) {
+		if (centerColumnIndex == currentCenterColumnIndex)
+		{
 			return;
 		}
 		currentCenterColumnIndex = centerColumnIndex;
@@ -625,16 +694,20 @@ public class HexGrid : MonoBehaviour {
 
 		Vector3 position;
 		position.y = position.z = 0f;
-		for (int i = 0; i < columns.Length; i++) {
-			if (i < minColumnIndex) {
+		for (int i = 0; i < columns.Length; i++)
+		{
+			if (i < minColumnIndex)
+			{
 				position.x = chunkCountX *
 					(HexMetrics.innerDiameter * HexMetrics.chunkSizeX);
 			}
-			else if (i > maxColumnIndex) {
+			else if (i > maxColumnIndex)
+			{
 				position.x = chunkCountX *
 					-(HexMetrics.innerDiameter * HexMetrics.chunkSizeX);
 			}
-			else {
+			else
+			{
 				position.x = 0f;
 			}
 			columns[i].localPosition = position;