+++ CHANGELOG.md
@@ -11,7 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - *HexFlags* enumeration type to represent bit flags for cell data.
 - Store road and river *HexCell* data in a bit flag field.
+- *HexCell.Grid* property, which is used to retrieve neighbor cells.
 - *HexCell.HasIncomingRiverThroughEdge* method.
+- *HexCell.TryGetNeighbor* and *HexGrid.TryGetCell* methods, to make cell retrieval more efficient and convenient.
+- *HexCoordinates.Step* method, to help retrieve cell neighbors based on coordinates.
 
 ### Changed
 
@@ -20,7 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
-- *HexCell* roads array.
+- *HexCell* neighbors and roads arrays.
 - *HexCell* incoming and outgoing river boolean and direction fields.
 - *HexCell.RiverBeginOrEndDirection* property.